refactor(transcript): split batch reduction context
