+++ scripts/max_lines_allowlist.txt
@@ -36,7 +36,7 @@ anyharness/crates/anyharness-lib/src/persistence/custom_migrations.rs 1126 custo
 anyharness/sdk-react/src/hooks/sessions.ts 723 existing SDK React oversized file
 cloud/sdk/src/types/generated.ts 796 existing generated cloud SDK type surface (+workspaceKind, git numstat fields, OpenAPI-derived materialization types, app-extra overlay, and PR 5 exact-ref create request fields)
 anyharness/sdk/src/reducer/__tests__/transcript.test.ts 1817 existing SDK oversized test file
-anyharness/sdk/src/reducer/transcript.ts 1542 existing SDK oversized file
+anyharness/sdk/src/reducer/transcript.ts 1529 existing SDK oversized file
 apps/desktop/src-tauri/src/commands/ssh_tunnel.rs 729 existing desktop native oversized file
 server/proliferate/auth/identity/store.py 686 managed sandbox GitHub grant read path extends existing identity store
 server/proliferate/config.py 618 managed Cloud, GitHub App, and Workflow flags extend the centralized deployment settings surface; split pending
